Allow ignore attrs for BenCharts
Allow user to specify attributes to ignore when making RunGroups. This
makes it possible to have a RunGroup in which more than one attribute
differs across Runs but they are in the same RunGroup -- making it so
later we can plot on the same Axes Runs with different values for
multiple attributes.