Fix data load to dataframe
Dataframe merging was not correct. Now, merge on the index which we set
for each datframe before merging them. Simplify the code to first prefix
then merge and finally make relative.

Also, pass utc=True to pd.to_datetime() so that if the timestamp doesn't
have a timezone it will add a utc timezone. This is what the docs say:

To do this, timezone-naive inputs are localized as UTC, while
timezone-aware inputs are converted to UTC.

Not sure exactly how this works, but...

I shouldn't include data that is timestamp without timezone tho